backend/requirements.txt
# FastAPI and Server
fastapi==0.108.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2

# LangChain and Related
langchain==0.1.0
langchain-anthropic==0.0.1
langchain-community==0.0.10
langchain-core==0.1.1
langchain-huggingface==0.0.1

# Vector Store and Embeddings
chromadb==0.4.15
sentence-transformers==2.2.2

# Utilities
numpy==1.24.3
pandas==2.0.3
python-jose[cryptography]==3.3.0
requests==2.31.0

# Development Tools
black==23.12.1
isort==5.13.2
pytest==7.4.3